requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
matplotlib==3.8.2
numpy==1.26.2
Pillow>=10.1.0
python-dateutil==2.8.2
